refactor: Reorganizar `input[type='range']`
